some cleanup. Refactor done. Dockerization TBD
