+++ .gitignore
@@ -2,6 +2,8 @@
 # Created by https://www.toptal.com/developers/gitignore/api/flask
 # Edit at https://www.toptal.com/developers/gitignore?templates=flask
 
+backup
+
 ### Flask ###
 instance/*
 !instance/.gitignore
