+++ app/routes.py
@@ -2,13 +2,6 @@ from flask import Blueprint, current_app, render_template, send_from_directory, 
 from config import DIRECTORY
 calendar = Blueprint("calendar", __name__)
 
-"""
-"/": sends the general calendar
-"<hash>": sends the custom hash calendar
-"/fetch_projects": api endpoint
-"/list": shows the projects to pick 
-"""
-
 @calendar.route("/")
 def get_general_calendar():
 	dir, calendar_file = current_app.calendar.create_full_calendar()
